+++ src/user.rs
@@ -16,7 +16,7 @@ pub type UserEventId = u16;
 /// Rules to extend some aspects of the battle with user defined behavior.
 pub trait UserRules<R: BattleRules> {
     /// See [UserMetricId](type.UserMetricId.html).
-    type UserMetricId: Eq + Hash + Clone + Debug;
+    type UserMetricId: Eq + Hash + Clone + Debug + Send;
     #[cfg(feature = "serialization")]
     /// See [UserEventPackage](type.UserEventPackage.html).
     type UserEventPackage: UserEventPacker<R>;
@@ -39,12 +39,12 @@ where
     /// Returns a boxed trait object version of this packed user event.
     ///
     /// Returns an error if the conversion failed.
-    fn boxed(self) -> WeaselResult<Box<dyn Event<R>>, R>;
+    fn boxed(self) -> WeaselResult<Box<dyn Event<R> + Send>, R>;
 
     /// Returns a UserEventPacker corresponding to the user event contained inside `event`.
     ///
     /// Fails if `event` is not an user event or if the conversion failed.
-    fn flattened(event: Box<dyn Event<R>>) -> WeaselResult<Self, R>;
+    fn flattened(event: Box<dyn Event<R> + Send>) -> WeaselResult<Self, R>;
 }
 
 #[cfg(feature = "serialization")]
@@ -52,13 +52,13 @@ impl<R> UserEventPacker<R> for ()
 where
     R: BattleRules,
 {
-    fn boxed(self) -> WeaselResult<Box<dyn Event<R>>, R> {
+    fn boxed(self) -> WeaselResult<Box<dyn Event<R> + Send>, R> {
         Err(WeaselError::UserEventUnpackingError(
             "empty UserEventPacker".into(),
         ))
     }
 
-    fn flattened(event: Box<dyn Event<R>>) -> WeaselResult<Self, R> {
+    fn flattened(event: Box<dyn Event<R> + Send>) -> WeaselResult<Self, R> {
         Err(WeaselError::UserEventPackingError(
             event.clone(),
             "empty UserEventPacker".into(),