+++ tests/client_server_test.rs
@@ -1,6 +1,5 @@
-use std::cell::RefCell;
 use std::ops::Range;
-use std::rc::Rc;
+use std::sync::{Arc, Mutex};
 use weasel::battle::{Battle, BattleRules};
 use weasel::entity::EntityId;
 use weasel::event::{
@@ -30,7 +29,7 @@ const PLAYER_2_ID: PlayerId = 2;
 /// Retrieves events from a server or client
 macro_rules! events {
     ($source: expr) => {{
-        $source.borrow().battle().history().events()
+        $source.lock().unwrap().battle().history().events()
     }};
 }
 
@@ -38,7 +37,8 @@ macro_rules! add_sink {
     ($source: expr, $sink: expr) => {{
         assert_eq!(
             $source
-                .borrow_mut()
+                .lock()
+                .unwrap()
                 .client_sinks_mut()
                 .add_sink(Box::new($sink.clone()))
                 .err(),
@@ -51,7 +51,8 @@ macro_rules! add_sink_from {
     ($source: expr, $sink: expr, $start: expr) => {{
         assert_eq!(
             $source
-                .borrow_mut()
+                .lock()
+                .unwrap()
                 .client_sinks_mut()
                 .add_sink_from(Box::new($sink.clone()), $start)
                 .err(),
@@ -81,14 +82,14 @@ impl SinkImpl {
 
 /// A test `ServerSink` sending events to a local server.
 struct TestServerSink<R: BattleRules> {
-    sink: Rc<RefCell<SinkImpl>>,
-    server: Rc<RefCell<Server<R>>>,
+    sink: Arc<Mutex<SinkImpl>>,
+    server: Arc<Mutex<Server<R>>>,
 }
 
 impl<R: BattleRules + 'static> TestServerSink<R> {
-    fn new(id: EventSinkId, server: Rc<RefCell<Server<R>>>) -> TestServerSink<R> {
+    fn new(id: EventSinkId, server: Arc<Mutex<Server<R>>>) -> TestServerSink<R> {
         TestServerSink {
-            sink: Rc::new(RefCell::new(SinkImpl::new(id))),
+            sink: Arc::new(Mutex::new(SinkImpl::new(id))),
             server,
         }
     }
@@ -105,46 +106,46 @@ impl<R: BattleRules> Clone for TestServerSink<R> {
 
 impl<R: BattleRules> EventSink for TestServerSink<R> {
     fn id(&self) -> EventSinkId {
-        self.sink.borrow().id
+        self.sink.lock().unwrap().id
     }
 
     fn on_disconnect(&mut self) {
-        self.sink.borrow_mut().disconnections += 1;
+        self.sink.lock().unwrap().disconnections += 1;
     }
 }
 
 impl<R: BattleRules + 'static> ServerSink<R> for TestServerSink<R> {
     fn send(&mut self, event: &ClientEventPrototype<R>) -> WeaselResult<(), R> {
-        if self.sink.borrow().broken {
+        if self.sink.lock().unwrap().broken {
             Err(WeaselError::EventSinkError("broken".to_string()))
         } else {
-            self.server.borrow_mut().process_client(event.clone())
+            self.server.lock().unwrap().process_client(event.clone())
         }
     }
 }
 
 /// A test `ClientSink` sending events to a local client.
 struct TestClientSink<R: BattleRules> {
-    sink: Rc<RefCell<SinkImpl>>,
-    client: Rc<RefCell<Client<R>>>,
-    buffer: Rc<RefCell<Vec<VersionedEventWrapper<R>>>>,
+    sink: Arc<Mutex<SinkImpl>>,
+    client: Arc<Mutex<Client<R>>>,
+    buffer: Arc<Mutex<Vec<VersionedEventWrapper<R>>>>,
 }
 
 impl<R: BattleRules + 'static> TestClientSink<R> {
-    fn new(id: EventSinkId, client: Rc<RefCell<Client<R>>>) -> TestClientSink<R> {
+    fn new(id: EventSinkId, client: Arc<Mutex<Client<R>>>) -> TestClientSink<R> {
         TestClientSink {
-            sink: Rc::new(RefCell::new(SinkImpl::new(id))),
+            sink: Arc::new(Mutex::new(SinkImpl::new(id))),
             client,
-            buffer: Rc::new(RefCell::new(Vec::new())),
+            buffer: Arc::new(Mutex::new(Vec::new())),
         }
     }
 
     /// Dumps all received events into the client.
     /// It's needed because it is not possible to borrow_mut client inside send().
     fn receive(&mut self) -> WeaselResult<(), R> {
-        let vec: Vec<_> = self.buffer.borrow_mut().drain(..).collect();
+        let vec: Vec<_> = self.buffer.lock().unwrap().drain(..).collect();
         for event in vec.into_iter() {
-            self.client.borrow_mut().receive(event)?;
+            self.client.lock().unwrap().receive(event)?;
         }
         Ok(())
     }
@@ -162,20 +163,20 @@ impl<R: BattleRules> Clone for TestClientSink<R> {
 
 impl<R: BattleRules> EventSink for TestClientSink<R> {
     fn id(&self) -> EventSinkId {
-        self.sink.borrow().id
+        self.sink.lock().unwrap().id
     }
 
     fn on_disconnect(&mut self) {
-        self.sink.borrow_mut().disconnections += 1;
+        self.sink.lock().unwrap().disconnections += 1;
     }
 }
 
 impl<R: BattleRules> ClientSink<R> for TestClientSink<R> {
     fn send(&mut self, event: &VersionedEventWrapper<R>) -> WeaselResult<(), R> {
-        if self.sink.borrow().broken {
+        if self.sink.lock().unwrap().broken {
             Err(WeaselError::EventSinkError("broken".to_string()))
         } else {
-            self.buffer.borrow_mut().push(event.clone());
+            self.buffer.lock().unwrap().push(event.clone());
             Ok(())
         }
     }
@@ -184,10 +185,10 @@ impl<R: BattleRules> ClientSink<R> for TestClientSink<R> {
 #[test]
 fn send_events() {
     // Create a server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Create a client.
-    let client = Rc::new(RefCell::new(util::client(CustomRules::new(), server_sink)));
+    let client = Arc::new(Mutex::new(util::client(CustomRules::new(), server_sink)));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink!(server, client_sink);
@@ -195,9 +196,9 @@ fn send_events() {
     let event_recorder = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink!(client, event_recorder);
     // Send an event from server and one from client.
-    util::team(&mut *server.borrow_mut(), TEAM_1_ID);
+    util::team(&mut *server.lock().unwrap(), TEAM_1_ID);
     assert_eq!(client_sink.receive().err(), None);
-    util::dummy(&mut *client.borrow_mut());
+    util::dummy(&mut *client.lock().unwrap());
     assert_eq!(client_sink.receive().err(), None);
     // Verify whether both battles have the same history.
     assert_eq!(events!(server)[0].kind(), EventKind::CreateTeam);
@@ -206,11 +207,11 @@ fn send_events() {
     assert_eq!(events!(client)[1].kind(), EventKind::DummyEvent);
     // Verify that the event recorder stores the same event as the client and server.
     assert_eq!(
-        event_recorder.buffer.borrow()[0].kind(),
+        event_recorder.buffer.lock().unwrap()[0].kind(),
         EventKind::CreateTeam
     );
     assert_eq!(
-        event_recorder.buffer.borrow()[1].kind(),
+        event_recorder.buffer.lock().unwrap()[1].kind(),
         EventKind::DummyEvent
     );
 }
@@ -218,14 +219,14 @@ fn send_events() {
 #[test]
 fn send_errors() {
     // Create a server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Create two clients.
-    let client_1 = Rc::new(RefCell::new(util::client(
+    let client_1 = Arc::new(Mutex::new(util::client(
         CustomRules::new(),
         server_sink.clone(),
     )));
-    let client_2 = Rc::new(RefCell::new(util::client(
+    let client_2 = Arc::new(Mutex::new(util::client(
         CustomRules::new(),
         server_sink.clone(),
     )));
@@ -235,8 +236,8 @@ fn send_errors() {
     let mut client_sink_2 = TestClientSink::new(CLIENT_2_ID, client_2.clone());
     add_sink!(server, client_sink_2);
     // Send event from one client. One client sink is faulty.
-    client_sink_2.sink.borrow_mut().broken = true;
-    util::dummy(&mut *client_1.borrow_mut());
+    client_sink_2.sink.lock().unwrap().broken = true;
+    util::dummy(&mut *client_1.lock().unwrap());
     assert_eq!(client_sink_1.receive().err(), None);
     assert_eq!(client_sink_2.receive().err(), None);
     // Event should be in the server and in one client.
@@ -244,12 +245,12 @@ fn send_errors() {
     assert_eq!(events!(client_1).len(), 1);
     assert_eq!(events!(client_2).len(), 0);
     // Check if the faulty sink got disconnected.
-    assert_eq!(client_sink_2.sink.borrow().disconnections, 1);
-    assert_eq!(server.borrow().client_sinks().sinks().count(), 1);
+    assert_eq!(client_sink_2.sink.lock().unwrap().disconnections, 1);
+    assert_eq!(server.lock().unwrap().client_sinks().sinks().count(), 1);
     // Server sink is faulty. Check no new event is added to the server.
-    server_sink.sink.borrow_mut().broken = true;
+    server_sink.sink.lock().unwrap().broken = true;
     assert_eq!(
-        DummyEvent::trigger(&mut *client_1.borrow_mut())
+        DummyEvent::trigger(&mut *client_1.lock().unwrap())
             .fire()
             .err(),
         Some(WeaselError::EventSinkError("broken".to_string()))
@@ -260,27 +261,28 @@ fn send_errors() {
 #[test]
 fn integrity_checks() {
     // Create a server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Create a client.
-    let client = Rc::new(RefCell::new(util::client(
+    let client = Arc::new(Mutex::new(util::client(
         CustomRules::new(),
         server_sink.clone(),
     )));
     // Fire event.
-    util::dummy(&mut *server.borrow_mut());
+    util::dummy(&mut *server.lock().unwrap());
     // Connect the client, but not from history start.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink!(server, client_sink);
     // Client should block the next server event.
-    util::dummy(&mut *server.borrow_mut());
+    util::dummy(&mut *server.lock().unwrap());
     assert_eq!(
         client_sink.receive().err(),
         Some(WeaselError::NonContiguousEventId(1, 0))
     );
     // Reattach client from history start.
     server
-        .borrow_mut()
+        .lock()
+        .unwrap()
         .client_sinks_mut()
         .remove_sink(CLIENT_1_ID);
     add_sink_from!(server, client_sink, 0);
@@ -288,27 +290,28 @@ fn integrity_checks() {
     assert_eq!(events!(server).len(), 2);
     assert_eq!(events!(client).len(), 2);
     // Client should receive the next event.
-    util::team(&mut *server.borrow_mut(), TEAM_1_ID);
-    util::creature(&mut *server.borrow_mut(), CREATURE_1_ID, TEAM_1_ID, ());
+    util::team(&mut *server.lock().unwrap(), TEAM_1_ID);
+    util::creature(&mut *server.lock().unwrap(), CREATURE_1_ID, TEAM_1_ID, ());
     assert_eq!(client_sink.receive().err(), None);
     assert_eq!(events!(server).len(), 4);
     assert_eq!(events!(client).len(), 4);
     // Fire an event in client. It should be processed correctly.
-    util::dummy(&mut *client.borrow_mut());
+    util::dummy(&mut *client.lock().unwrap());
     assert_eq!(client_sink.receive().err(), None);
     assert_eq!(events!(server).len(), 5);
     assert_eq!(events!(client).len(), 5);
     // Change server.
-    assert_eq!(server_sink.sink.borrow().disconnections, 0);
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    assert_eq!(server_sink.sink.lock().unwrap().disconnections, 0);
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let new_server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     client
-        .borrow_mut()
+        .lock()
+        .unwrap()
         .set_server_sink(Box::new(new_server_sink));
-    assert_eq!(server_sink.sink.borrow().disconnections, 1);
+    assert_eq!(server_sink.sink.lock().unwrap().disconnections, 1);
     // Fire another event in the client.
     assert_eq!(
-        StartRound::trigger(&mut *client.borrow_mut(), ENTITY_1_ID)
+        StartRound::trigger(&mut *client.lock().unwrap(), ENTITY_1_ID)
             .fire()
             .err(),
         Some(WeaselError::EntityNotFound(ENTITY_1_ID))
@@ -325,24 +328,26 @@ fn check_version() {
     // Create a server with newer rules.
     let mut rules = CustomRules::new();
     rules.version = VERSION_NEW;
-    let server = Rc::new(RefCell::new(util::server(rules)));
+    let server = Arc::new(Mutex::new(util::server(rules)));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Create a client with older rules.
     let mut rules = CustomRules::new();
     rules.version = VERSION_OLD;
-    let client = Rc::new(RefCell::new(util::client(rules, server_sink)));
+    let client = Arc::new(Mutex::new(util::client(rules, server_sink)));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink_from!(server, client_sink, 0);
     // Check if events from server are rejected.
-    util::dummy(&mut *server.borrow_mut());
+    util::dummy(&mut *server.lock().unwrap());
     assert_eq!(
         client_sink.receive().err(),
         Some(WeaselError::IncompatibleVersions(VERSION_OLD, VERSION_NEW))
     );
     // Check if events from client are rejected.
     assert_eq!(
-        DummyEvent::trigger(&mut *client.borrow_mut()).fire().err(),
+        DummyEvent::trigger(&mut *client.lock().unwrap())
+            .fire()
+            .err(),
         Some(WeaselError::IncompatibleVersions(VERSION_OLD, VERSION_NEW))
     );
 }
@@ -350,21 +355,22 @@ fn check_version() {
 #[test]
 fn add_client_sink() {
     // Create server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Fire four events.
     for _ in 0..4 {
-        util::dummy(&mut *server.borrow_mut());
+        util::dummy(&mut *server.lock().unwrap());
     }
     assert_eq!(events!(server).len(), 4);
     // Create client.
-    let client = Rc::new(RefCell::new(util::client(CustomRules::new(), server_sink)));
+    let client = Arc::new(Mutex::new(util::client(CustomRules::new(), server_sink)));
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     // Add client sink with invalid range.
     let range = Range { start: 5, end: 7 };
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .client_sinks_mut()
             .add_sink_range(Box::new(client_sink.clone()), range.clone())
             .err(),
@@ -373,7 +379,8 @@ fn add_client_sink() {
     let range = Range { start: 0, end: 7 };
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .client_sinks_mut()
             .add_sink_range(Box::new(client_sink.clone()), range.clone())
             .err(),
@@ -382,7 +389,8 @@ fn add_client_sink() {
     // Add the client sink and send the first two events.
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .client_sinks_mut()
             .add_sink_range(Box::new(client_sink.clone()), Range { start: 0, end: 2 })
             .err(),
@@ -393,7 +401,8 @@ fn add_client_sink() {
     // Check id is verified when sending events.
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .client_sinks_mut()
             .send_range(CLIENT_ERR_ID, Range { start: 0, end: 2 })
             .err(),
@@ -402,7 +411,8 @@ fn add_client_sink() {
     // Send the other two events.
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .client_sinks_mut()
             .send_range(CLIENT_1_ID, Range { start: 2, end: 4 })
             .err(),
@@ -418,7 +428,7 @@ fn rights() {
     let server = Server::builder(Battle::builder(CustomRules::new()).build())
         .enforce_authentication()
         .build();
-    let server = Rc::new(RefCell::new(server));
+    let server = Arc::new(Mutex::new(server));
     // Create a client with auth.
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     let client = Client::builder(
@@ -427,27 +437,29 @@ fn rights() {
     )
     .enable_authentication(PLAYER_1_ID)
     .build();
-    let client = Rc::new(RefCell::new(client));
-    assert_eq!(client.borrow().authentication(), true);
+    let client = Arc::new(Mutex::new(client));
+    assert_eq!(client.lock().unwrap().authentication(), true);
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink!(server, client_sink);
     // Check that new rights for non existing team are rejected.
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .rights_mut()
             .add(PLAYER_1_ID, &TEAM_1_ID)
             .err(),
         Some(WeaselError::TeamNotFound(TEAM_1_ID))
     );
     // Create a team and some rights.
-    util::team(&mut *server.borrow_mut(), TEAM_1_ID);
-    util::creature(&mut *server.borrow_mut(), CREATURE_1_ID, TEAM_1_ID, ());
+    util::team(&mut *server.lock().unwrap(), TEAM_1_ID);
+    util::creature(&mut *server.lock().unwrap(), CREATURE_1_ID, TEAM_1_ID, ());
     assert_eq!(client_sink.receive().err(), None);
     assert_eq!(
         server
-            .borrow_mut()
+            .lock()
+            .unwrap()
             .rights_mut()
             .add(PLAYER_2_ID, &TEAM_1_ID)
             .err(),
@@ -455,7 +467,7 @@ fn rights() {
     );
     // Check that rights are enforced for the wrong client.
     assert_eq!(
-        StartRound::trigger(&mut *client.borrow_mut(), ENTITY_1_ID)
+        StartRound::trigger(&mut *client.lock().unwrap(), ENTITY_1_ID)
             .fire()
             .err(),
         Some(WeaselError::AuthenticationError(
@@ -465,18 +477,19 @@ fn rights() {
     );
     // Create a client without any authentication.
     server
-        .borrow_mut()
+        .lock()
+        .unwrap()
         .client_sinks_mut()
         .remove_sink(CLIENT_1_ID);
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
-    let client = Rc::new(RefCell::new(util::client(CustomRules::new(), server_sink)));
+    let client = Arc::new(Mutex::new(util::client(CustomRules::new(), server_sink)));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink_from!(server, client_sink, 0);
     assert_eq!(client_sink.receive().err(), None);
     // Client events should be rejected.
     assert_eq!(
-        StartRound::trigger(&mut *client.borrow_mut(), ENTITY_1_ID)
+        StartRound::trigger(&mut *client.lock().unwrap(), ENTITY_1_ID)
             .fire()
             .err(),
         Some(WeaselError::MissingAuthentication)
@@ -489,14 +502,14 @@ fn rights() {
     )
     .enable_authentication(PLAYER_2_ID)
     .build();
-    let client = Rc::new(RefCell::new(client));
+    let client = Arc::new(Mutex::new(client));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_2_ID, client.clone());
     add_sink_from!(server, client_sink, 0);
     assert_eq!(client_sink.receive().err(), None);
     // Check that the good client can send events.
     assert_eq!(
-        StartRound::trigger(&mut *client.borrow_mut(), ENTITY_1_ID)
+        StartRound::trigger(&mut *client.lock().unwrap(), ENTITY_1_ID)
             .fire()
             .err(),
         None,
@@ -506,9 +519,9 @@ fn rights() {
 #[test]
 fn server_only_events() {
     // Create a client and a server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let mut server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
-    let client = Rc::new(RefCell::new(util::client(
+    let client = Arc::new(Mutex::new(util::client(
         CustomRules::new(),
         server_sink.clone(),
     )));
@@ -517,13 +530,13 @@ fn server_only_events() {
     add_sink!(server, client_sink);
     // Verify that client blocks server-only events.
     assert_eq!(
-        CreateTeam::trigger(&mut *client.borrow_mut(), TEAM_1_ID)
+        CreateTeam::trigger(&mut *client.lock().unwrap(), TEAM_1_ID)
             .fire()
             .err(),
         Some(WeaselError::ServerOnlyEvent)
     );
     // Verify that server blocks server-only events from clients.
-    let event = CreateTeam::trigger(&mut *client.borrow_mut(), TEAM_1_ID)
+    let event = CreateTeam::trigger(&mut *client.lock().unwrap(), TEAM_1_ID)
         .prototype()
         .client_prototype(0, None);
     assert_eq!(
@@ -540,19 +553,19 @@ fn client_server_serde() {
 
     const ENTITY_1_ID: EntityId<CustomRules> = EntityId::Creature(CREATURE_1_ID);
     // Create a server.
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
     // Create a client.
-    let client = Rc::new(RefCell::new(util::client(CustomRules::new(), server_sink)));
+    let client = Arc::new(Mutex::new(util::client(CustomRules::new(), server_sink)));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
     add_sink!(server, client_sink);
     // Send events from server and from client.
-    util::team(&mut *server.borrow_mut(), TEAM_1_ID);
-    util::creature(&mut *server.borrow_mut(), TEAM_1_ID, CREATURE_1_ID, ());
+    util::team(&mut *server.lock().unwrap(), TEAM_1_ID);
+    util::creature(&mut *server.lock().unwrap(), TEAM_1_ID, CREATURE_1_ID, ());
     assert_eq!(client_sink.receive().err(), None);
     assert_eq!(
-        StartRound::trigger(&mut *client.borrow_mut(), ENTITY_1_ID)
+        StartRound::trigger(&mut *client.lock().unwrap(), ENTITY_1_ID)
             .fire()
             .err(),
         None
@@ -576,24 +589,24 @@ fn client_server_serde() {
         ]
     );
     // Start a new server and load history.
-    let history_json = helper::history_as_json(server.borrow().battle());
-    let server = Rc::new(RefCell::new(util::server(CustomRules::new())));
+    let history_json = helper::history_as_json(server.lock().unwrap().battle());
+    let server = Arc::new(Mutex::new(util::server(CustomRules::new())));
     let server_sink = TestServerSink::new(SERVER_1_ID, server.clone());
-    helper::load_json_history(&mut *server.borrow_mut(), history_json);
+    helper::load_json_history(&mut *server.lock().unwrap(), history_json);
     // Start a new client and load history.
-    let history_json = helper::history_as_json(server.borrow().battle());
-    let client = Rc::new(RefCell::new(util::client(CustomRules::new(), server_sink)));
+    let history_json = helper::history_as_json(server.lock().unwrap().battle());
+    let client = Arc::new(Mutex::new(util::client(CustomRules::new(), server_sink)));
     // Connect the client to the server.
     let mut client_sink = TestClientSink::new(CLIENT_1_ID, client.clone());
-    helper::load_json_history(&mut *client.borrow_mut(), history_json);
+    helper::load_json_history(&mut *client.lock().unwrap(), history_json);
     add_sink!(server, client_sink);
     // Fire new events.
     assert_eq!(
-        EndRound::trigger(&mut *client.borrow_mut()).fire().err(),
+        EndRound::trigger(&mut *client.lock().unwrap()).fire().err(),
         None
     );
     assert_eq!(
-        RemoveCreature::trigger(&mut *server.borrow_mut(), CREATURE_1_ID)
+        RemoveCreature::trigger(&mut *server.lock().unwrap(), CREATURE_1_ID)
             .fire()
             .err(),
         None